drm/xe: Only use delayed worker to destroy userptr VMAs
TTM has a delayed BO destruction mechanism based on dma-resv, which
eliminates the need for the delayed worker to destroy VMAs containing
BOs. However, userptr still requires this mechanism because DMA mappings
and notifiers cannot be removed until the GPU bindings are gone.
Therefore, move the delayed worker responsible for destroying userptr
VMAs to the userptr code and bypass this process for VMAs containing a
BO.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>